Add Weather - AerisWeather
AerisWeather is a small weather data company that prides itself on being friendly to developers and the development community. We found your list from a post on reddit and love all the resources it aggregates. We really hope your readers enjoy our documentation and our free tier to create really cool products.